+++ doc.go
@@ -1,3 +1,3 @@
 // Package httpmux is an HTTP request multiplexer
-// that allows mux nesting.
+// that allows router nesting and middleware stacking.
 package httpmux